Fix lazy property init
